Update post_gen_project.py
More debugging

+++ hooks/post_gen_project.py
@@ -71,11 +71,18 @@ def copy_os_specific_readme():
     os_type = 'windows' if platform.system() == 'Windows' else 'macos'
 
     # Define the paths to the source and destination README.md files
-    src_readme_path = os.path.join('os_specific_files', f'README_{os_type}.md')
+    src_readme_path = os.path.abspath(os.path.join('os_specific_files', f'README_{os_type}.md'))
     print(f"Src README: {src_readme_path}")
-    dest_readme_path = os.path.join('{{cookiecutter.project_slug}}', 'README.md')
+    dest_readme_path = os.path.abspath(os.path.join('{{cookiecutter.project_slug}}', 'README.md'))
     print(f"Dest README: {dest_readme_path}")
-    
+
+    if not os.path.exists(src_readme_path):
+        print(f"Source file does not exist: {src_readme_path}")
+        return
+    if not os.path.exists(os.path.dirname(dest_readme_path)):
+        print(f"Destination directory does not exist: {os.path.dirname(dest_readme_path)}")
+        return
+
     # Copy the appropriate README.md file to the project root
     shutil.copy(src_readme_path, dest_readme_path)
